requirements.txt
# Core dependencies (REQUIRED)
requests>=2.28.0
PyYAML>=6.0

# Optional dependencies for specific backends
# Uncomment the ones you need:

# For Kafka support:
# kafka-python>=2.0.2

# For AWS CloudWatch support:
# boto3>=1.26.0

# For GCP Cloud Logging support:
# google-cloud-logging>=3.5.0

# For Azure Monitor support:
# (uses built-in requests, no extra dependency)

# Development and testing
# pytest>=7.0.0
# pytest-cov>=4.0.0
# flake8>=6.0.0
# pylint>=2.17.0
